renames files & classes for arm
